feat: renamed ParameterErrors to singular name + added jsdoc
