+++ lib/ParameterValidator.js
@@ -1,9 +1,20 @@
 import debug from 'debug';
 import inspector from 'schema-inspector';
 import _ from 'lodash';
-import ParameterError from './errors/ParameterErrors.js';
+import ParameterError from './errors/ParameterError.js';
 
+/**
+ * Parameter Validation class (should be initiated once per endpoint upon its setup)
+ * @module ParameterValidator
+ */
 export default class ParameterValidator {
+  /**
+   * ParameterValidator
+   * @property {Object[]} parametersDefinition - openapi parameters definition
+   * @see {@link https://swagger.io/docs/specification/describing-parameters/} - for parametersDefinition
+   * @property {Object}   requestBodyDefinition - openapi body definition
+   * @see {@link https://swagger.io/docs/specification/describing-request-body/} - for requestBodyDefinition
+   */
   constructor(parametersDefinition, requestBodyDefinition) {
     const pathParamsDefinition = parametersDefinition?.filter(({ in: source }) => source === 'path');
     const qsParamsDefinition = parametersDefinition?.filter(({ in: source }) => source === 'query');
@@ -13,6 +24,14 @@ export default class ParameterValidator {
     this.setupSchema();
   }
 
+  /**
+   * Parameter builder from openapi definition to schema-inspector
+   * @private
+   * @param {object} parent parent object in which we need to edit its properties
+   * @param {integer} i     current iterator
+   * @return {object}
+   * @see {@link https://www.npmjs.com/package/schema-inspector} - for returned result
+   */
   static buildParam(parent, i) {
     const param = parent.properties[i];
     const correlations = _.pick(param, ['type', 'minLength', 'maxLength', 'properties', 'required', 'items']);
@@ -47,6 +66,11 @@ export default class ParameterValidator {
     return newParam;
   }
 
+  /**
+   * Recursion class that goes through each open api definition (to convert to schema-inspector)
+   * @private
+   * @param {object} param
+   */
   eachRecursive(param) {
     // eslint-disable-next-line no-param-reassign
     if (param.properties) {
@@ -61,6 +85,10 @@ export default class ParameterValidator {
     }
   }
 
+  /**
+   * Setup the dynamic schema of parameters (path / query / body)
+   * @private
+   */
   setupSchema() {
     this.schema = {
       type: 'object',
@@ -116,6 +144,13 @@ export default class ParameterValidator {
     this.debug('created validation schema', JSON.stringify(this.schema));
   }
 
+  /**
+   * test an incoming request against the current instance of ParameterValidator
+   * @param {Map<string, any>} path
+   * @param {Map<string, any>} qs
+   * @param {Map<string, any>} requestBody
+   * @throws module:ParameterError
+   */
   test(path = {}, qs = {}, requestBody = {}) {
     this.debug('testing parameter', path, qs, requestBody);
     // TODO: check string byte / binary via custom validator
